fix: 🐛 make Point.lat/lng float -> int

+++ src/main/java/org/meteogroup/pointcompression/Point.java
@@ -2,27 +2,27 @@ package org.meteogroup.pointcompression;
 
 public class Point {
 
-  private float lat;
-  private float lon;
+  private int lat;
+  private int lon;
 
-  public Point(float lat, float lon) {
+  public Point(int lat, int lon) {
     this.lat = lat;
     this.lon = lon;
   }
 
-  public float getLat() {
+  public int getLat() {
     return lat;
   }
 
-  public void setLat(float lat) {
+  public void setLat(int lat) {
     this.lat = lat;
   }
 
-  public float getLon() {
+  public int getLon() {
     return lon;
   }
 
-  public void setLon(float lon) {
+  public void setLon(int lon) {
     this.lon = lon;
   }
